# Refactoring router receiver