--- a/.github/CONTRIBUTING.md
+++ b/.github/CONTRIBUTING.md
@@ -4,6 +4,5 @@
 2. Create a feature branch
 3. Commit your changes
 4. Rebase your local changes against the master branch
-5. Run test suite with the `go test ./...` command and confirm that it passes
-6. Run `gofmt -s`
-7. Create new Pull Request
+5. Fix your codes if CI failed.
+6. Create new Pull Request