--- a/.github/ISSUE_TEMPLATE/bug_report.md
+++ b/.github/ISSUE_TEMPLATE/bug_report.md
@@ -4,32 +4,23 @@ about: Create a report to help us improve
 
 ---
 
-**Describe the bug**
-A clear and concise description of what the bug is.
+# Description
 
-**To Reproduce**
-Steps to reproduce the behavior:
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
+# Expected Behavior
 
-**Expected behavior**
-A clear and concise description of what you expected to happen.
+# Current Behavior
 
-**Screenshots**
-If applicable, add screenshots to help explain your problem.
+# Steps to Reproduce
+1.
+2.
+3.
+4.
 
-**Desktop (please complete the following information):**
- - OS: [e.g. iOS]
- - Browser [e.g. chrome, safari]
- - Version [e.g. 22]
+# Environment
+<!--- Tell us your go version. -->
 
-**Smartphone (please complete the following information):**
- - Device: [e.g. iPhone6]
- - OS: [e.g. iOS8.1]
- - Browser [e.g. stock browser, safari]
- - Version [e.g. 22]
+# Possible Solution
+<!--- Not obligatory, but tell us if you have any ideas. -->
 
-**Additional context**
-Add any other context about the problem here.
+# Possible Implementation
+<!--- Not obligatory, but tell us if you have any ideas. -->