--- a/.github/ISSUE_TEMPLATE/question.md
+++ b/.github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,7 @@
+---
+name: Question
+about: Create a report to ask any quetions.
+
+---
+
+# Question