--- a/.github/PULL_REQUEST_TEMPLATE.md
+++ b/.github/PULL_REQUEST_TEMPLATE.md
@@ -1,4 +1,4 @@
-# Overview
+# Description
 
 # Changes
 
@@ -7,5 +7,7 @@
 # Operational Requirements
 
 # Related Issue
+<!--- Not obligatory, but write any issues if exists it -->
 
 # Supplement
+<!--- Not obligatory -->