# Overview
Improve perfomance by stopped using Router pointer receiver.

# Changes
- Stopped using Router pointer receiver
- Add benchstat tool
- Fix bugs ing CI jobs
  - Required jobs were not running in go1.19 
- End of support go1.15
- Update issue templates

Before
```sh
go test -bench=. -cpu=1 -benchmem -count=3
goos: darwin
goarch: amd64
pkg: github.com/bmf-san/goblin
cpu: VirtualApple @ 2.50GHz
BenchmarkStatic1         9382904               128.2 ns/op            32 B/op          1 allocs/op
BenchmarkStatic1         9410863               127.6 ns/op            32 B/op          1 allocs/op
BenchmarkStatic1         9417586               126.6 ns/op            32 B/op          1 allocs/op
BenchmarkStatic5         4367858               274.8 ns/op            96 B/op          1 allocs/op
BenchmarkStatic5         4368003               286.1 ns/op            96 B/op          1 allocs/op
BenchmarkStatic5         4197505               276.0 ns/op            96 B/op          1 allocs/op
BenchmarkStatic10        2445247               496.3 ns/op           176 B/op          1 allocs/op
BenchmarkStatic10        2535044               480.3 ns/op           176 B/op          1 allocs/op
BenchmarkStatic10        2482634               470.7 ns/op           176 B/op          1 allocs/op
BenchmarkWildCard1       1874241               640.1 ns/op           384 B/op          6 allocs/op
BenchmarkWildCard1       1738502               703.2 ns/op           384 B/op          6 allocs/op
BenchmarkWildCard1       1851855               719.1 ns/op           384 B/op          6 allocs/op
BenchmarkWildCard5        507542              2472 ns/op             928 B/op         13 allocs/op
BenchmarkWildCard5        486055              2470 ns/op             928 B/op         13 allocs/op
BenchmarkWildCard5        522135              2443 ns/op             928 B/op         13 allocs/op
BenchmarkWildCard10       259974              4631 ns/op            1560 B/op         19 allocs/op
BenchmarkWildCard10       256924              4629 ns/op            1560 B/op         19 allocs/op
BenchmarkWildCard10       261069              4609 ns/op            1560 B/op         19 allocs/op
BenchmarkRegexp1         1963184               615.9 ns/op           384 B/op          6 allocs/op
BenchmarkRegexp1         1950442               611.6 ns/op           384 B/op          6 allocs/op
BenchmarkRegexp1         1970739               613.0 ns/op           384 B/op          6 allocs/op
BenchmarkRegexp5          533304              2309 ns/op             928 B/op         13 allocs/op
BenchmarkRegexp5          534800              2326 ns/op             928 B/op         13 allocs/op
BenchmarkRegexp5          517088              2325 ns/op             928 B/op         13 allocs/op
BenchmarkRegexp10         289285              4329 ns/op            1560 B/op         19 allocs/op
BenchmarkRegexp10         276583              4330 ns/op            1560 B/op         19 allocs/op
BenchmarkRegexp10         272575              4527 ns/op            1560 B/op         19 allocs/op
PASS
ok      github.com/bmf-san/goblin       40.444s
```

After
```sh
go test -bench=. -cpu=1 -benchmem -count=3
goos: darwin
goarch: amd64
pkg: github.com/bmf-san/goblin
cpu: VirtualApple @ 2.50GHz
BenchmarkStatic1        10072951               123.4 ns/op            32 B/op          1 allocs/op
BenchmarkStatic1        10089045               119.7 ns/op            32 B/op          1 allocs/op
BenchmarkStatic1         9869005               119.5 ns/op            32 B/op          1 allocs/op
BenchmarkStatic5         4678380               255.9 ns/op            96 B/op          1 allocs/op
BenchmarkStatic5         4672666               257.4 ns/op            96 B/op          1 allocs/op
BenchmarkStatic5         4618848               253.0 ns/op            96 B/op          1 allocs/op
BenchmarkStatic10        2847400               427.5 ns/op           176 B/op          1 allocs/op
BenchmarkStatic10        2854250               422.3 ns/op           176 B/op          1 allocs/op
BenchmarkStatic10        2722155               434.5 ns/op           176 B/op          1 allocs/op
BenchmarkWildCard1       1903887               632.0 ns/op           384 B/op          6 allocs/op
BenchmarkWildCard1       1888086               626.7 ns/op           384 B/op          6 allocs/op
BenchmarkWildCard1       1914217               633.9 ns/op           384 B/op          6 allocs/op
BenchmarkWildCard5        500449              2394 ns/op             928 B/op         13 allocs/op
BenchmarkWildCard5        533448              2486 ns/op             928 B/op         13 allocs/op
BenchmarkWildCard5        487317              2464 ns/op             928 B/op         13 allocs/op
BenchmarkWildCard10       271914              4541 ns/op            1560 B/op         19 allocs/op
BenchmarkWildCard10       266788              4601 ns/op            1560 B/op         19 allocs/op
BenchmarkWildCard10       260335              4581 ns/op            1560 B/op         19 allocs/op
BenchmarkRegexp1         1887534               616.6 ns/op           384 B/op          6 allocs/op
BenchmarkRegexp1         1947963               612.8 ns/op           384 B/op          6 allocs/op
BenchmarkRegexp1         1959728               605.0 ns/op           384 B/op          6 allocs/op
BenchmarkRegexp5          517082              2300 ns/op             928 B/op         13 allocs/op
BenchmarkRegexp5          538726              2302 ns/op             928 B/op         13 allocs/op
BenchmarkRegexp5          522044              2307 ns/op             928 B/op         13 allocs/op
BenchmarkRegexp10         278421              4287 ns/op            1560 B/op         19 allocs/op
BenchmarkRegexp10         287530              4325 ns/op            1560 B/op         19 allocs/op
BenchmarkRegexp10         281001              4393 ns/op            1560 B/op         19 allocs/op
PASS
ok      github.com/bmf-san/goblin       39.821s
```

# Impact range
Method using Router's pointer receiver.

# Operational Requirements
N/A

# Related Issue
https://github.com/bmf-san/goblin/issues/70

# Supplement
N/A